Renamed the bst_list_view dir to just bst.
Files checked in: DataRepo/tests/static/js/tests.html DataRepo/tests/views/models/bst_list_view/column/filterer/test_annotation.py DataRepo/tests/views/models/bst_list_view/column/filterer/test_field.py DataRepo/tests/views/models/bst_list_view/column/sorter/test_annotation.py DataRepo/tests/views/models/bst_list_view/column/sorter/test_field.py DataRepo/tests/views/models/bst_list_view/column/sorter/test_many_related_field.py DataRepo/tests/views/models/bst_list_view/column/test_annotation.py DataRepo/tests/views/models/bst_list_view/column/test_field.py DataRepo/tests/views/models/bst_list_view/column/test_many_related_field.py DataRepo/tests/views/models/bst_list_view/column/test_many_related_group.py DataRepo/tests/views/models/bst_list_view/column/test_related_field.py DataRepo/tests/views/models/bst_list_view/test_base.py DataRepo/tests/views/models/bst_list_view/test_client_interface.py DataRepo/views/models/bst_list_view/__init__.py DataRepo/views/models/bst_list_view/base.py DataRepo/views/models/bst_list_view/client_interface.py DataRepo/views/models/bst_list_view/column/__init__.py DataRepo/views/models/bst_list_view/column/annotation.py DataRepo/views/models/bst_list_view/column/base.py DataRepo/views/models/bst_list_view/column/field.py DataRepo/views/models/bst_list_view/column/filterer/__init__.py DataRepo/views/models/bst_list_view/column/filterer/annotation.py DataRepo/views/models/bst_list_view/column/filterer/base.py DataRepo/views/models/bst_list_view/column/filterer/field.py DataRepo/views/models/bst_list_view/column/many_related_field.py DataRepo/views/models/bst_list_view/column/many_related_group.py DataRepo/views/models/bst_list_view/column/related_field.py DataRepo/views/models/bst_list_view/column/sorter/__init__.py DataRepo/views/models/bst_list_view/column/sorter/annotation.py DataRepo/views/models/bst_list_view/column/sorter/base.py DataRepo/views/models/bst_list_view/column/sorter/field.py DataRepo/views/models/bst_list_view/column/sorter/many_related_field.py DataRepo/views/models/bst/__init__.py DataRepo/views/models/bst/base.py DataRepo/views/models/bst/client_interface.py DataRepo/views/models/bst/column/__init__.py DataRepo/views/models/bst/column/annotation.py DataRepo/views/models/bst/column/base.py DataRepo/views/models/bst/column/field.py DataRepo/views/models/bst/column/filterer/__init__.py DataRepo/views/models/bst/column/filterer/annotation.py DataRepo/views/models/bst/column/filterer/base.py DataRepo/views/models/bst/column/filterer/field.py DataRepo/views/models/bst/column/many_related_field.py DataRepo/views/models/bst/column/many_related_group.py DataRepo/views/models/bst/column/related_field.py DataRepo/views/models/bst/column/sorter/__init__.py DataRepo/views/models/bst/column/sorter/annotation.py DataRepo/views/models/bst/column/sorter/base.py DataRepo/views/models/bst/column/sorter/field.py DataRepo/views/models/bst/column/sorter/many_related_field.py
