+++ DataRepo/tests/static/js/tests.html
@@ -53,15 +53,15 @@
         -->
 
         <!-- Value Tests -->
-        <script src="/DataRepo/static/js/bst_list_view/value.js"></script>
+        <script src="/DataRepo/static/js/bst/value.js"></script>
         <script src="bst_list_view/test_value.js"></script>
 
         <!-- Sorter Tests -->
-        <script src="/DataRepo/static/js/bst_list_view/sorter.js"></script>
+        <script src="/DataRepo/static/js/bst/sorter.js"></script>
         <script src="bst_list_view/test_sorter.js"></script>
 
         <!-- Filterer Tests -->
-        <script src="/DataRepo/static/js/bst_list_view/filterer.js"></script>
+        <script src="/DataRepo/static/js/bst/filterer.js"></script>
         <script src="bst_list_view/test_filterer.js"></script>
 
     </body>
