+++ DataRepo/tests/views/models/bst_list_view/column/filterer/test_annotation.py
@@ -3,7 +3,7 @@ from django.test import override_settings
 
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 from DataRepo.utils.exceptions import DeveloperWarning
-from DataRepo.views.models.bst_list_view.column.filterer.annotation import (
+from DataRepo.views.models.bst.column.filterer.annotation import (
     BSTAnnotFilterer,
 )
