+++ DataRepo/tests/views/models/bst_list_view/column/sorter/test_field.py
@@ -8,7 +8,7 @@ from DataRepo.tests.tracebase_test_case import (
     create_test_model,
 )
 from DataRepo.utils.exceptions import DeveloperWarning
-from DataRepo.views.models.bst_list_view.column.sorter.field import BSTSorter
+from DataRepo.views.models.bst.column.sorter.field import BSTSorter
 
 BSTSTestModel = create_test_model(
     "BSTSTestModel",
