+++ DataRepo/tests/views/models/bst_list_view/column/sorter/test_many_related_field.py
@@ -15,7 +15,7 @@ from DataRepo.tests.tracebase_test_case import (
     create_test_model,
 )
 from DataRepo.utils.exceptions import DeveloperWarning
-from DataRepo.views.models.bst_list_view.column.sorter.many_related_field import (
+from DataRepo.views.models.bst.column.sorter.many_related_field import (
     BSTManyRelatedSorter,
 )
